refactor: create useFilters hook
